flows: split agent-only flow definitions out of production.py
production.py keeps invite_intake, post_meeting and live_meeting — the three
flows that still DO something with no agent domain. production_agent.py takes
meeting_prep, email_chat, desk_setup and desk_claim, which degrade to nothing,
and registers them only when common.domain_present("agent") — the same
predicate every needs=("agent",) step already consults.

_their_clock and _scaffold_refs are promoted from build() closures to module
level: both halves mint scaffolds, and a recipe with two homes is the defect
_note_path exists to close. CLAIM_BOOK moves with await_claim, its only reader.
